Query rustc about the actual crate name
rather than trying to find it by manually parsing Rust code looking for any `#![crate_name]`s!
This is a lot more robust and more correct! We're now able to correctly deduce the actual
crate name even if e.g. `--crate-name` was passed verbatim via directives or on the command line.

It does drop the detection of `#![crate_type]` for which there is no good replacement since there
isn't an equivalent `--print=crate-types`. In practice that means we no longer add `--extern proc_macro`
when the actual crate type is specified via the attribute. I think this is absolute fine.
As a replacment, in the future, we'll make it easier to set the crate type to proc-macro on the command line.