Default build modes: Imply unstable options in most cases
Previously, we'd only imply it if the user supplied a
rrx-native flag / one tracked by us of which we obviously
knew if they unstable or not.

I'm tired of having to add it in cases where I'm providing
unstable verbatim args. E.g.,
`rrd … -- --doctest-compilation-args … -Zunstable-options`,
`--persist-doctests=…` and so on.