Fix `//@ no-prefer-dynamic` "for good"
The gift that keeps on giving.
Fixes regression introduced in 47d943852d3f5d78c67c10ed036c6670a8d3779e.
We were defaulting to a crate type of lib for principal
(e.g, non-auxiliary) files which is crazily incorrect.
At the time of writing, principal tests ought to default to rustc's
default (which is bin).